src/web/requirements.txt
opencv-python
numpy
pillow
pyzbar